Extract SeriesList component and fix margin mismatch
