+++ src/components/PodcastSeriesList.js
@@ -1,34 +1,22 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { ScrollView, StyleSheet, View } from 'react-native';
 
+import SeriesList from './SeriesList';
 import PodcastSeriesTile from './PodcastSeriesTile';
 import Podcast from '../state/models/Podcast';
 
-const styles = StyleSheet.create({
-  list: {
-    flex: 1,
-    flexDirection: 'row',
-    flexWrap: 'wrap',
-    justifyContent: 'flex-start',
-    paddingVertical: 7,
-  },
-});
+const PodcastSeriesList = ({ podcasts, onPressPodcast }) => (
+  <SeriesList>
+    {podcasts.map(podcast => (
+      <PodcastSeriesTile
+        key={podcast.title}
+        podcast={podcast}
+        onPress={() => onPressPodcast(podcast)}
+      />
+    ))}
+  </SeriesList>
+);
 
-const PodcastSeriesList = ({ podcasts, onPressPodcast }) =>
-  (
-    <ScrollView>
-      <View style={styles.list}>
-        {podcasts.map(podcast => (
-          <PodcastSeriesTile
-            key={podcast.title}
-            podcast={podcast}
-            onPress={() => onPressPodcast(podcast)}
-          />
-        ))}
-      </View>
-    </ScrollView>
-  );
 PodcastSeriesList.propTypes = {
   podcasts: PropTypes.arrayOf(
     PropTypes.shape(
